requirements.txt
requests==2.28.2 
gitpython==3.1.31
python-dotenv==0.20.0     